+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - increased timeout on finetune_llm_evaluation project from 1 hour (default) to 4 hours
 - pin `ray-operator`, `ray-cluster`, and `ray-image` modules versions
 - pin module versions for all manifests
+- the `sagemaker/sagemaker-model-package-promote-pipeline` module no longer generates a Docker image
 
 ## v1.4.0